[FEATURE] Remove default header output from backend previews
This removes the default preview header of each content element to make custom previews more flexible. In v12 the subheader was added as a fourth “fixed” text to every element preview, making custom previews more complicated (if the subeader field was used for something else).
We include header/subheader in the default preview Layout template, so updating should not affect most instances where current previews rely on the header field of an element to be displayed “by default”.